UpdatePayment

+++ src/main/java/dao/OrderItemDAO.java
@@ -9,8 +9,8 @@ import java.sql.Types;
 
 public class OrderItemDAO extends DBConnection {
 
-   
     public void addOrderItem(OrderItem item, Connection conn) throws SQLException {
+        // Cập nhật câu lệnh SQL để khớp với các trường trong model đã tối ưu
         String sql = "INSERT INTO dbo.OrderItems (OrderID, TicketInfoID, EventID, TicketID, UnitPrice, Quantity, TotalPrice, CreatedAt) VALUES (?, ?, ?, ?, ?, ?, ?, GETDATE())";
         
         try (PreparedStatement ps = conn.prepareStatement(sql)) {
@@ -21,15 +21,15 @@ public class OrderItemDAO extends DBConnection {
             if (item.getTicketId() != 0) {
                 ps.setInt(4, item.getTicketId());
             } else {
-                ps.setNull(4, Types.INTEGER); // Cho phép ticketId là null
+                ps.setNull(4, Types.INTEGER);
             }
             
-            ps.setDouble(5, item.getUnitPrice());
+            // SỬA LỖI Ở ĐÂY: Dùng setBigDecimal thay vì setDouble
+            ps.setBigDecimal(5, item.getUnitPrice());
             ps.setInt(6, item.getQuantity());
-            ps.setDouble(7, item.getTotalPrice());
+            ps.setBigDecimal(7, item.getTotalPrice());
             
             ps.executeUpdate();
         }
     }
-    
 }
